% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass Identification Data Set}
\format{A data frame with 214 rows and 10 variables:
\describe{
  \item{RI}{Refractive Index}
  \item{Na}{Sodium (unit measurement: weight percent in corresponding oxide, as are attributes 4-10)}
  \item{Mg}{Magnesium}
  \item{Al}{Aluminum}
  \item{Si}{Silicon}
  \item{K}{Potassium}
  \item{Ca}{Calcium}
  \item{Ba}{Barium}
  \item{Fe}{Iron}
  \item{Type}{Class Attribute}
  
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/glass+identification}
}
\usage{
glass
}
\description{
A dataset containing 6 types of glass; defined in terms of their oxide content
}
\keyword{datasets}
